- added change password and profile edit views - modified profile view to show more member information - added views to display reviews grouped by series, author with sorting - added series rating to reviews - added review comments - added scheduled TVDB update sync - modified TVDB syncs to use english translations - updated to Spring Boot 4.0.3 - schema changes - (dev) added config profiles

+++ src/main/java/com/kwsni/caught_up/tvdb/repository/SeriesRepository.java
@@ -8,6 +8,6 @@ import org.springframework.transaction.annotation.Transactional;
 import com.kwsni.caught_up.tvdb.model.Series;
 
 @Transactional
-public interface SeriesRepository extends JpaRepository<Series, Integer> {
+public interface SeriesRepository extends JpaRepository<Series, Long> {
     public Optional<Series> findBySlug(String slug);
 }
